BFD: Labeling a manually configured neighbor for tracking
For BFD controlled Static routes, the BFD sessions are created
to the route next hop. However, in some scenarios, e.g. when the
next hop itself does not support BFD, or when it might be more
meaningful to create the session to a remote monitoring station,
we need more flexible relationship between a static route nexthop
and its controlling BFD session.

In this patch, a numeric label value can be specified with manually
configured BFD neighbors, and a static route nexthop can be
configured to track a certain BFD session using that label. The
detailed behavior is updated in the user's guide.

This feature may also be added to other protocols, but it seems not
so meaningful.